requirements.txt
Flask
langchain
langchain_chroma
langchain_community
python-dotenv
langchain_core
llama-cpp-python
sentence-transformers
pandas
scikit-learn